+++ DesignToolbox/DesignToolboxUITests/Screenshots/Documentation Screenshots - Controls.swift
@@ -14,6 +14,7 @@
 import XCTest
 
 // swiftlint:disable required_deinit
+// swiftlint:disable type_name
 
 /// Walks across the application and make screenshots for documentation.
 ///
@@ -47,7 +48,7 @@ import XCTest
 /// Screenshots can be found in tests attachements.
 /// You can pick them in tests reports.
 /// Then get the screenshots in attachments ; base name of the files are the ones defined in the tests.
-final class ControlsDocumentationScreenshots: AppTestCase {
+final class 📄_ControlsDocumentationScreenshots: AppTestCase {
 
     // MARK: - Checkboxes
 
@@ -298,3 +299,4 @@ final class ControlsDocumentationScreenshots: AppTestCase {
 }
 
 // swiftlint:enable required_deinit
+// swiftlint:enable type_name
