+++ DesignToolbox/DesignToolboxUITests/Screenshots/Documentation Screenshots - Indicators.swift
@@ -14,6 +14,7 @@
 import XCTest
 
 // swiftlint:disable required_deinit
+// swiftlint:disable type_name
 
 /// Walks across the application and make screenshots for documentation.
 ///
@@ -38,7 +39,7 @@ import XCTest
 /// Screenshots can be found in tests attachements.
 /// You can pick them in tests reports.
 /// Then get the screenshots in attachments ; base name of the files are the ones defined in the tests.
-final class IndicatorsDocumentationScreenshots: AppTestCase {
+final class 📄_IndicatorsDocumentationScreenshots: AppTestCase {
 
     // MARK: - Badge
 
@@ -88,3 +89,4 @@ final class IndicatorsDocumentationScreenshots: AppTestCase {
 }
 
 // swiftlint:enable required_deinit
+// swiftlint:enable type_name
